+++ backends/xla/ops.go
@@ -325,46 +325,6 @@ func (f *Function) Conj(operand backends.Value) (backends.Value, error) {
 	return f.newNode(value), nil
 }
 
-// Dot returns the "dot product" operation.
-// The exact semantics of this operation depend on the ranks of the operands:
-// | Input | Output | Semantics |
-// | vector [n] dot vector [n] | scalar | vector dot product |
-// | matrix [m x k] dot vector [k] | vector [m]	matrix-vector multiplication |
-// | matrix [m x k] dot matrix [k x n] | matrix [m x n] | matrix-matrix multiplication |
-// The operation performs sum of products over the second dimension of x0 (or the first if it has rank 1) and
-// the first dimension of x1.
-// These are the "contracted" dimensions.
-// The contracted dimensions of x0 and x1 must be of the same size.
-// In practice, it can be used to perform dot products between vectors, vector/matrix multiplications, or
-// matrix/matrix multiplications.
-func (f *Function) Dot(lhs, rhs backends.Value) (backends.Value, error) {
-	nodes, err := f.verifyAndCastValues("Dot", lhs, rhs)
-	if err != nil {
-		return nil, err
-	}
-	lhsNode, rhsNode := nodes[0], nodes[1]
-	var output backends.Value
-	if lhsNode.shape.Rank() == 1 && rhsNode.shape.Rank() == 1 {
-		// Contracting both vectors.
-		output, err = f.DotGeneral(lhsNode, []int{0}, []int{}, rhsNode, []int{0}, []int{})
-	} else if lhsNode.shape.Rank() == 2 && rhsNode.shape.Rank() == 1 {
-		// Contract rhs vector.
-		output, err = f.DotGeneral(lhsNode, []int{1}, []int{}, rhsNode, []int{0}, []int{})
-	} else if lhsNode.shape.Rank() == 1 && rhsNode.shape.Rank() == 2 {
-		// Contract lhs vector.
-		output, err = f.DotGeneral(lhsNode, []int{0}, []int{}, rhsNode, []int{0}, []int{})
-	} else if lhsNode.shape.Rank() == 2 && rhsNode.shape.Rank() == 2 {
-		// Traditional matrix multiplication:
-		output, err = f.DotGeneral(lhsNode, []int{1}, []int{}, rhsNode, []int{0}, []int{})
-	} else {
-		return nil, errors.Errorf("Dot operands have invalid ranks: lhs=%v, rhs=%v", lhsNode.shape, rhsNode.shape)
-	}
-	if err != nil {
-		return nil, errors.WithMessagef(err, "while building op Dot()")
-	}
-	return output, nil
-}
-
 // Clamp returns the element-wise clamping operation.
 //
 // The values max and min can either be a scalar or have the same shape as x.